Remove atom_ and bitvect_private.h
Moves functionality inside the C file.
